boot loader phase refactoring and SPL support
Some boards (e.g. cubox, arnadale, etc) support the booloader to be embedded
inside the first megabyte of the bootdevice.

The support for this configuration  is triggered only if a bootloader
(or a series of bootloaders) is defined in the BOOTLOADERS field but no BOOTDIR
is specified in LAYOUT.
The syntax for this new feature is:

uboot-prefix: myboard
bootloaders: SPL:1 u-boot.img:42

that tranlates into:

dd if=ROOTDIRFS/usr/lib/u-boot/myboard/SPL of=/dev/loopX bs=1K seek=1
dd if=ROOTDIRFS/usr/lib/u-boot/myboard/u-boot.img of=/dev/loopX bs=1K seek=42

While here, add the possibility to specify an uEnv.txt file per board and do
some cleanup.

Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>